split rpad and lpad
